requirements.txt
nose>=1.3.7
urllib3==1.22
numpy>=1.15.4
numpydoc>=0.8.0,<0.9.0
absl-py>=0.7.0
tqdm>=4.31.1
matplotlib>=2.2,<3.1
joblib>=0.13.2
dpath>=1.4.2
mlflow>=1.0.0
networkx>=2.3
